add loading + server to add team
